pyproject.toml
[tool.poetry]
name = "factory"
version = "0.0.0a"
description = "General Purpose G(enerative)AI service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
diffusers = "^0.27.2"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
torch = "^2.2.2"
accelerate = "^0.29.1"
transformers = "^4.39.3"
peft = "^0.10.0"
torchao = "^0.1"
opencv-python = "^4.9.0.80"
insightface = "^0.7.3"
ip-adapterv = "^0.1.0"
einops = "^0.7.0"
torchvision = "^0.17.2"
pillow = "^10.3.0"
onnxruntime = "^1.17.3"
numpy = "^1.22.0"
pydub = "^0.25.1"
bark = "^0.1.5"
nltk = "^3.8.1"
qrcode = "^7.4.2"

[tool.poetry.group.dev.dependencies]
pillow = "^10.3.0"
numpy = "^1.26.4"
datasets = "^2.18.0"
tqdm = "^4.66.2"
torchvision = "^0.17.2"
jupyter = "^1.0.0"
notebook = "^7.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
